renamed space exploration feat
